% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{geGetColors}
\alias{geGetColors}
\title{geGetColors}
\usage{
geGetColors(n, alpha = 1)
}
\arguments{
\item{n}{the number of colors to use (1-17)}

\item{alpha}{the alpha value for colors [default: 1]}
}
\value{
a vector of colors in hex code
}
\description{
geGetColors
}
