% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_reduction_axes.R
\name{geom_reduction_axes_label}
\alias{geom_reduction_axes_label}
\title{geom_reduction_axes_label}
\usage{
geom_reduction_axes_label(
  mapping = NULL,
  len = 0.1,
  offset = 0.1,
  label = "X",
  axis = 1,
  ...
)
}
\description{
geom_reduction_axes_label
}
