% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_set.r
\name{geUseTheme}
\alias{geUseTheme}
\title{geUseTheme}
\usage{
geUseTheme(theme = "bw")
}
\arguments{
\item{theme}{ggplot themes (either "bw" or "classic")}
}
\description{
Set the theme in the current session
}
