% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.r
\name{italicize.plot}
\alias{italicize.plot}
\title{italicize.plot}
\usage{
italicize.plot(p, string)
}
\arguments{
\item{p}{a ggplot object}

\item{string}{string or vector of strings to replace}
}
\description{
italicize everything that matches the `string` in a given plot `p`
}
\examples{
library(ggessentials)
data(iris)
p <- ggplot(iris, aes(x=Species, y=Sepal.Width, color=Species)) +
  geom_boxplot()
g <- italicize.plot(p, c('setosa','versicolor'))
plot(g)

}
\references{
https://github.com/GuangchuangYu/yyplot/blob/master/R/set_font.R
}
