% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.r
\name{italicize.string}
\alias{italicize.string}
\title{italicize.string}
\usage{
italicize.string(sentence, taxon, ...)
}
\arguments{
\item{sentence}{a sentence}

\item{taxon}{a string to italicize}
}
\description{
italicize everything that matches the `taxon` in a given sentence
}
