% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{scale_fill_ge}
\alias{scale_fill_ge}
\title{scale_fill_ge}
\usage{
scale_fill_ge(primary = "blue", other = "grey", direction = 1, ...)
}
\arguments{
\item{primary}{the color used for the primary case [default: "blue"]}

\item{other}{the color used for the other case [default: "grey"]}

\item{direction}{positive for normal order, negative for reversed order [default: 1 - normal order]}

\item{...}{other parameters passed to `ggplot2::discrete_scale`}
}
\description{
scale_fill_ge
}
