% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_reduction_axes.R
\docType{data}
\name{StatReductionAxes}
\alias{StatReductionAxes}
\title{StatReductionAxes}
\format{
An object of class \code{StatReductionAxes} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatReductionAxes
}
\description{
StatReductionAxes
}
\keyword{datasets}
