% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_reduction_axes.R
\docType{data}
\name{StatReductionAxesLabel}
\alias{StatReductionAxesLabel}
\title{StatReductionAxesLabel}
\format{
An object of class \code{StatReductionAxesLabel} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatReductionAxesLabel
}
\description{
StatReductionAxesLabel
}
\keyword{datasets}
