% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_set.r
\name{theme_umap_ge}
\alias{theme_umap_ge}
\title{theme_umap_ge}
\usage{
theme_umap_ge(aes_x = NULL, aes_y = NULL, vjust.x = 2, vjust.y = -1)
}
\arguments{
\item{aes_x}{for patchwork object generated by DimPlot, specify the reduction axis, e.g. aes_x="UMAP_1"}

\item{aes_y}{for patchwork object generated by DimPlot, specify the reduction axis, e.g. aes_y="UMAP_2"}

\item{vjust.x}{vertical adjustment for label on x axis}

\item{vjust.y}{vertical adjustment for label on y axis}
}
\description{
theme_umap_ge
}
