% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{geGetColors}
\alias{geGetColors}
\title{geGetColors}
\usage{
geGetColors(n = 17, alpha = 1, named = FALSE)
}
\arguments{
\item{n}{the number of colors to use (1-17) [default: 17]}

\item{alpha}{the alpha value for colors [default: 1]}

\item{named}{return a list with color names [default: FALSE]}
}
\value{
a vector of colors in hex code
}
\description{
geGetColors
}
