% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.r
\name{italicize.plot}
\alias{italicize.plot}
\title{italicize.plot}
\usage{
italicize.plot(p, string)
}
\arguments{
\item{p}{a ggplot object}

\item{string}{string to replace}
}
\description{
italicize everything that matches the `string` in a given plot `p`
}
